fix(upload_files): fixed bugs with reference file deletion
- temp files are now removed correctly when final files are uploaded
- deleting files removes mod connection first and then removes the file if there are no other mods linked
