+++ agr_literature_service/api/routers/referencefile_router.py
@@ -138,7 +138,7 @@ def delete(referencefile_id: int,
            user: OktaUser = db_user,
            db: Session = db_session):
     set_global_user_from_okta(db, user)
-    referencefile_crud.destroy(db, referencefile_id)
+    referencefile_crud.destroy(db, referencefile_id, get_okta_mod_access(user))
     return Response(status_code=status.HTTP_204_NO_CONTENT)
